Change Rack application to Sinatra
In preparation for implementing some actions in Ruby, we added the
Sinatra web framework to the `Gemfile`, re-bundled, and changed
`config.ru` to load a Sinatra application in app.rb.

At this point, `rake tomcat:run` still works to run the Spring MVC
app, but the `petclinic.war` won't deploy unless you run the
application server in an environment with `GEM_HOME` set to where the
gems were installed. It's possible to make a standalone war with the
gems packaged in it but we won't get to that until later.

We're also using the `sinatra-reloader` gem for code reloading, since
developing in Ruby doesn't make much sense if we can't change the code
on the fly. Code reloading is activated by `RACK_ENV=development`, which
we set by default in the `tomcat:run` rake task.

To experiment with code reloading, start Tomcat,
[view the main Sinatra page](http://localhost:3000/rack/),
change app.rb, and then reload the page. You should see your edits
take effect.