Implement `/rack/vets.xml`.
Next step in refactoring: Make the test green again.

In this commit we added a Sinatra `get` action for `/rack/vets.xml`
that renders the xml with Builder.

We also added some Spring helpers that use JRuby's Java integration to
look up the Spring context from within the servlet context, and make
it possible to get the Clinic object out of it for use by our Sinatra
action.