Move Builder logic to Sinatra views
As a last step in the "small" refactoring, we configure Sinatra to use
views and move the builder template logic to them. This greatly
simplifies the actions in `app.rb`, showing that it pays to keep view
logic separate from controller logic for all but the most trivial of
applications.

The JSON action could also be simplified through the use of a helper
module, but we'll leave that as an exercise.

We'll see the separation of concerns play more heavily as we move
toward Rails in the "medium" refactoring.