Change from a Rack servlet to a filter
This allows us to intercept every request and give Ruby a chance to
handle it. For this we're introducing a custom filter called
`RubyFirstRackFilter`, which is similar in operation to the `RackFilter`
that comes with JRuby-Rack except that Ruby gets first dibs on
handling the request. If Ruby returns an error, then control is turned
over to the rest of the application.

Note: The `RubyFirstRackFilter` will probably get merged into a future
release of JRuby-Rack. For now just consider it a part of the
library that you wouldn't need to maintain in your application.