Revert links back to pre-`/rack` versions and disable Spring content negotiation
Next, we're going to verify that we can implement actions in Ruby
mixed among the URI-space of application that still take precedence.
We'll do this by first inducing failures in the Cucumber suite, and
then rectifying them by changing Ruby code.

We're reverting the URL changes in c63937b9 and d3a937a2 to remove the
`/rack` prefixes and disabling the Spring content negotiation
configuration that serves up XML and Atom content. This causes three
of our extended features to fail:

    Failing Scenarios:
    cucumber -p extended features/vets.feature:3 # Scenario: View vets as XML
    cucumber -p extended features/vets.feature:10 # Scenario: View vets as JSON
    cucumber -p extended features/visits.feature:12 # Scenario: Visits Atom Feed

    14 scenarios (3 failed, 11 passed)
    79 steps (3 failed, 3 skipped, 73 passed)