Change from Sinatra to Rails
Now we're going to change from Sinatra to Rails. To do this, we first
generated an empty Rails app as follows:

    rails new spring-petclinic -O
    mv spring-petclinic/{app,config} src/main/webapp/WEB-INF
    rm -rf spring-petclinic

We're not quite doing full-blown Rails development yet, so we don't
need supporting directories like `script`, `test`, and `vendor`. (We
already have a `lib` directory under `WEB-INF`.)

We'll update our `Gemfile` to use Rails instead of Sinatra and
re-bundle.

We'll change web.xml to use JRuby-Rack's Rails context listener, which
provides some extra integration to the servlet environment
(connecting the Rails logger to the container, for example). This also
means we don't need `WEB-INF/config.ru` any more.

We don't need our Sinatra `app.rb` either, but we'll leave it for now
so that we can migrate the code to Rails.

At this point some of our Cucumber features are failing again. `rake`
now show:

    Failing Scenarios:
    cucumber features/vets.feature:3 # Scenario: View vets as XML
    cucumber features/visits.feature:12 # Scenario: Visits Atom Feed

That's to be expected, as we haven't implemented those actions yet.
Let's do that next.