Copy JSP contents to layout and new welcome view
For the next step in this refactoring, we're going to start serving
HTML views in Rails and transition over some of the JSPs to Rails
views.

In this incremental, incomplete step, we're copying the content of
`WEB-INF/jsps/welcome.jsp` to `app/views/welcome/index.html.erb` and
concatenating the content of the `includes.jsp`, `headers.jsp`, and
`footers.jsp` inside of `app/views/layouts/application.html.erb`.

We haven't converted the JSP content over to ERB yet because we want
to demonstrate that separately in the next commit.