Convert JSP-isms to ordinary URLs + ERb
In this step we convert all JSP tags in the layout and welcome view to
ordinary URLs. Next we'll wire up a route to a controller for the
welcome page.

We also put a customized banner graphic in the layout to visually
distinguish pages served by Rails from those still served by Spring.