Implement `owners#search` action
Here we're writing just the minimum amount of code to make the
features succeed again. To do that we need to add a route, a
controller, and ERB-ify the view.

What's interesting at this point is that the display of the search
form happens in Rails, but the processing of the form still happens in
Java/Spring. You can see this in the application at this point by
seeing that the Rails logo appears on the search form page, but once
you submit the form, it goes away.