Add empty Rails application at the root of the project
We're now going to embark on the "large" refactoring. What we consider
makes it "large" is that it's potentially more disruptive to your
existing Java development workflow. We'll aim to run the entire
application inside a `rails server` process. Changing the project
structure to mirror a regular Rails application allows us to use
generators and other Rails tools.

This commit is also a bit bigger because we'll need a couple more
things in place before we can resume running the Cucumber features,
Then, once we do, we'll immediately be staring at a bunch of failures
because we haven't implemented the rest of the application yet.