Move Rails code, assets and config into place
We're reclaiming the Rails code we already wrote and moving it to the
proper place in `app/` and `lib/`. We also moved images and
stylesheets, as well as helper code.

The application still doesn't really run, because there's no Spring
context -- we need to boot it ourselves, since there's no container to
do it for us. That Spring configuration will come in the next commit.