Implement `owner#search` action
We now can display a listing of owners in response to submitting the
search form. We're starting down a road of doing things more
Rails-like by adding an `Owner` model. Note that we're actually just
declaring the `Owner` Java object as our model, and decorating it with
some ruby mehods.

We get another scenario green but still have quite a few
to go.

    13 scenarios (9 failed, 4 passed)
    75 steps (9 failed, 25 skipped, 41 passed)