Add `owner#show`
The most interesting part of this commit is that we're starting to
turn the `Owner` hibernate model into an `ActiveModel`. By adding
`ActiveModel::Naming` and `ActiveModel::Conversion` and implementing
`Owner.name` and `Owner#persisted?` we get a `#to_param` for free
which allows the named route method `owner_path(owner_obj)` to work as
expected.

Two more scenarios pass.

    13 scenarios (7 failed, 6 passed)
    75 steps (7 failed, 18 skipped, 50 passed)