Get Spring context booting
To get the Spring context booting inside Rails, we have to do a couple
things.

1. We use Spring's `ClassPathXmlApplicationContext` class to boot the
   context, and add the `config` directory to the classpath so the
   Spring XML files can be found by it.
2. We have to ensure all the Spring jar files and dependencies are
   added to the classpath.

Normally this second step would be a huge pain if not for a nifty new
Maven plugin that ships with JRuby 1.6. Basically, you run this plugin
and it ensures you have all your jars downloaded and cached, and then
it spits out a Ruby script to add those jars to the classpath. You can
see the output in `config/initializers/classpath.rb`. There's a new
Rake task `maven_classpath` that can be used to (re-)generate the file.