Implement the visits atom feed
Here we switched the visits atom feed  to use Rails' `#atom_feed`
helper. We also switched the `visits#index` action to use a MIME
`#respond_to` to use content negotiation. As a result we no longer
need the custom route for the atom feed.

Now all the scenarios pass again.

    13 scenarios (13 passed)
    75 steps (75 passed)