Add a "View as JSON" link to the vets JSP
Now we're at a point where we can start to add new functionality to
the app. Here, we'll add a JSON version of the vets XML document.

Since this is a new feature not present in the original version of the
petclinic, we've included the feature to verify this in the `extended`
cucumber profile.

Running `rake cucumber:extended` at this point shows you some red:

    (::) failed steps (::)

    expected: /json/,
         got: "text/html;charset=utf-8" (using =~)
    Diff:
    @@ -1,2 +1,2 @@
    -/json/
    +text/html;charset=utf-8
     (RSpec::Expectations::ExpectationNotMetError)
    org/jruby/RubyProc.java:268:in `call'
    ./features/step_definitions/xml_json_steps.rb:21:in `(root)':in `/^I should see a JSON document$/'
    features/vets.feature:13:in `Then I should see a JSON document'

    Failing Scenarios:
    cucumber -p extended features/vets.feature:10 # Scenario: View vets as JSON

    14 scenarios (1 failed, 13 passed)
    79 steps (1 failed, 1 skipped, 77 passed)